async check_pypi_version
